requirements.txt
tensorflow
numpy
pillow
openai 
gradio